+++ frappe/integrations/doctype/google_contacts/google_contacts.json
@@ -1,16 +1,11 @@
 {
+ "autoname": "format:GC-{user}",
  "creation": "2019-06-11 23:47:40.278888",
  "doctype": "DocType",
  "engine": "InnoDB",
  "field_order": [
   "enable",
-  "user",
-  "google_credentials",
-  "client_id",
-  "client_secret",
-  "allow_contacts_access",
-  "authorization_code",
-  "refresh_token"
+  "user"
  ],
  "fields": [
   {
@@ -25,39 +20,9 @@
    "label": "User",
    "options": "User",
    "read_only": 1
-  },
-  {
-   "fieldname": "google_credentials",
-   "fieldtype": "Section Break",
-   "label": "Google Credentials"
-  },
-  {
-   "fieldname": "client_id",
-   "fieldtype": "Data",
-   "label": "Client ID"
-  },
-  {
-   "fieldname": "client_secret",
-   "fieldtype": "Data",
-   "label": "Client Secret"
-  },
-  {
-   "fieldname": "allow_contacts_access",
-   "fieldtype": "Button",
-   "label": "Allow Contacts Access"
-  },
-  {
-   "fieldname": "authorization_code",
-   "fieldtype": "Data",
-   "label": "Authorization Code"
-  },
-  {
-   "fieldname": "refresh_token",
-   "fieldtype": "Data",
-   "label": "Refresh Token"
   }
  ],
- "modified": "2019-06-11 23:48:41.889272",
+ "modified": "2019-06-12 00:40:32.392200",
  "modified_by": "Administrator",
  "module": "Integrations",
  "name": "Google Contacts",
