+++ frappe/integrations/doctype/google_contacts_settings/test_google_contacts_settings.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2019, Frappe Technologies and Contributors
+# See license.txt
+from __future__ import unicode_literals
+
+# import frappe
+import unittest
+
+class TestGoogleContactsSettings(unittest.TestCase):
+	pass
